reminder/requirements.txt
requests
schedule
python-dotenv
